requirements.txt
### 35. requirements.txt (Updated with all dependencies)
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
WTForms==3.1.1
bcrypt==4.1.2
pyserial==3.5
pynmea2==1.18.0
adif-io==0.1.2
requests==2.31.0
pyopenssl==23.3.0
numpy==1.24.3